add multi-monitor support (#134)
* rebase add multi-monitor support

https://github.com/yuka-friends/Windrecorder/pull/67

* update poetry

* clean wintitle: remove asterisk for saved state

* remove half res scale record setting

* add ui setting WIP

* add ui setting WIP

* add ui setting

* remove compress preset from config class

* add record encoder WIP

* add record encoder

* add multi and single display record

* add ocr mask process

* fix ocr crop img

* fix record fallback logic

* remove pyautogui in flag

* fix bug and improve perform

* add onboard setting

* add onboard setting

* tweak onboard setting

* add i18n

* add crop URBL fallback

* remove PyAutoGUI

* update poetry dependence

* add strategy condition in screenshot compare

* record base on bitrate not crf

* change setting to bitrate and add estimate

* Update languages.json

* remove disable when under auto index

* Update CHANGELOG.md

* Fix record strategy condition in set_display function

* update changelog

* completed edge options

* Update languages.json

* Add support for SVT-AV1 encoder and extract i-frames using FFmpeg

* Add PaddleOCR module installation script

* Add PaddleOCR support for text recognition

* Update logger wr.log using UTF8

* Update OCR logic thresholds

* Revert "Add PaddleOCR support for text recognition"

This reverts commit 93dc90a2d2610abfda5869c3d2e5b9a0bb10cc38.

* Revert "Add PaddleOCR module installation script"

This reverts commit c6c293cf1d5fad88ff2636d6c0bb375967d7c567.

* revert default record method due to performance compatibility

* add ocr img preprocessor

* update onboard setting

* fix display record setting not appoint config

* update i18n and polish setting

* tweak setting

* fix find_closest_video_by_filesys if empty list

* tweak i18n

* Update CHANGELOG.md

* Update CHANGELOG.md

* update img

---------

+++ windrecorder/file_utils.py
@@ -14,12 +14,17 @@ logger = get_logger(__name__)
 
 # 清空指定目录下的所有文件和子目录
 def empty_directory(path):
+    if len(path) == 0:
+        return
     with os.scandir(path) as it:
         for entry in it:
-            if entry.is_dir():
-                shutil.rmtree(entry.path)
-            else:
-                os.remove(entry.path)
+            try:
+                if entry.is_dir():
+                    shutil.rmtree(entry.path)
+                else:
+                    os.remove(entry.path)
+            except Exception as e:
+                logger.error(e)
 
 
 # 检查目录是否存在，若无则创建
